buildroot reusable workflow
I switched the rootfs.yml to being a reusable workflow and also enabled caching of the output when it's run as a submodule with the reusable workflow. This means that it can skip the actual build and just reuse the cached one in most cases, moving the total time for the job down under a minute!